# specs/001-medical-triage-nav/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TriNav Medical Triage API
  description: |
    Medical triage and hospital navigation assistant.

    **Safety Notice**: This system provides triage guidance ONLY, not medical diagnosis.
    Always seek professional medical care for emergencies.

    **Compliance**: Health information service (non-diagnostic) per Chinese regulations.
  version: 1.0.0
  contact:
    name: TriNav API Support
    email: support@trinav.example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.trinav.example.com
    description: Production environment

tags:
  - name: triage
    description: Medical triage operations

paths:
  /assistant/invoke:
    post:
      tags:
        - triage
      summary: Submit symptoms and receive triage assessment
      description: |
        Main entry point for medical triage workflow.

        **Input**: Text description (required), optional image, optional GPS coordinates
        **Output**: Triage level, recommended departments, possible causes, self-care tips, red flags, navigation (if GPS provided)

        **Performance**: EMERGENCY ≤5s, ROUTINE ≤15s
      operationId: triage.invoke
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - text
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: Unique conversation identifier (generate new UUID if not provided)
                  example: "550e8400-e29b-41d4-a716-446655440000"
                text:
                  type: string
                  maxLength: 2000
                  description: Symptom description in Chinese
                  example: "手臂出现红疹，有点痒，持续2天"
                image_base64:
                  type: string
                  format: byte
                  description: Base64-encoded image (JPEG/PNG, ≤4MB after compression)
                  example: "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBD..."
                lat:
                  type: number
                  format: float
                  minimum: -90
                  maximum: 90
                  description: Latitude for hospital navigation
                  example: 39.9042
                lng:
                  type: number
                  format: float
                  minimum: -180
                  maximum: 180
                  description: Longitude for hospital navigation
                  example: 116.4074
            required:
              - session_id
              - text
      responses:
        '200':
          description: Successful triage assessment
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [need_more_info, final, error]
                    description: Workflow status
                  session_id:
                    type: string
                    format: uuid
                    description: Session identifier
                  turn_count:
                    type: integer
                    minimum: 1
                    maximum: 2
                    description: Current clarification round
                  # === Output when status=final ===
                  triage_level:
                    type: string
                    enum: [EMERGENCY, URGENT, ROUTINE, SELF_CARE]
                    description: Urgency assessment (present when status=final)
                  recommended_departments:
                    type: array
                    items:
                      type: string
                    description: Suggested medical departments (present when status=final)
                    example: ["皮肤科"]
                  possible_causes:
                    type: array
                    items:
                      type: string
                    maxItems: 3
                    description: Suspected causes with qualified language (present when status=final)
                    example: ["过敏相关皮疹（疑似）", "接触性皮炎（疑似）"]
                  self_care_tips:
                    type: array
                    items:
                      type: string
                    maxItems: 10
                    description: Non-prescriptive care guidance (present when status=final)
                    example: ["避免抓挠", "记录皮疹变化"]
                  red_flags:
                    type: array
                    items:
                      type: string
                    maxItems: 10
                    description: Warning signs requiring emergency care (present when status=final)
                    example: ["如果出现呼吸困难/脸唇肿胀/全身迅速扩散，请立刻急诊"]
                  # === Navigation (optional, if GPS provided) ===
                  navigation:
                    type: object
                    description: Hospital recommendations (present when GPS provided and status=final)
                    properties:
                      hospitals:
                        type: array
                        items:
                          type: object
                          properties:
                            rank:
                              type: integer
                              minimum: 1
                              maximum: 3
                            name:
                              type: string
                            is_3a:
                              type: boolean
                            distance_m:
                              type: integer
                            reason:
                              type: string
                          required:
                            - rank
                            - name
                            - is_3a
                            - reason
                        minItems: 3
                        maxItems: 3
                      route_plan:
                        type: object
                        properties:
                          mode:
                            type: string
                            enum: [driving, transit, walking]
                          eta_min:
                            type: integer
                          summary:
                            type: string
                        required:
                          - mode
                          - eta_min
                          - summary
                  # === Weather (optional) ===
                  weather:
                    type: object
                    description: Weather-related travel tips (present if available)
                    properties:
                      summary:
                        type: string
                      tips:
                        type: array
                        items:
                          type: string
                  # === Mandatory fields ===
                  disclaimer:
                    type: string
                    description: Legal disclaimer (present when status=final)
                    example: "本建议仅供参考，不替代专业医疗诊断"
                  hotline_tip:
                    type: string
                    description: Contact fallback (present when status=final)
                    example: "如果你不确定症状严重程度，或者情况在变重，建议你也可以拨打当地医疗热线或直接拨打医院电话先确认"
                  # === Output when status=need_more_info ===
                  clarify_questions:
                    type: array
                    items:
                      type: string
                    maxItems: 3
                    description: Clarification questions (present when status=need_more_info)
                    example: ["是否伴有发热?", "疼痛部位具体在哪里?", "是否有恶心呕吐?"]
                  # === Error handling ===
                  error_message:
                    type: string
                    description: Error details (present when status=error)
                    example: "External service temporarily unavailable"
                required:
                  - status
                  - session_id
                  - disclaimer
                  - hotline_tip
              examples:
                routine_triage_with_navigation:
                  summary: ROUTINE triage with hospital navigation
                  value:
                    status: "final"
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    turn_count: 1
                    triage_level: "ROUTINE"
                    recommended_departments: ["皮肤科"]
                    possible_causes:
                      - "过敏相关皮疹（疑似）"
                      - "接触性皮炎（疑似）"
                    self_care_tips:
                      - "避免抓挠患处"
                      - "记录皮疹变化"
                    red_flags:
                      - "如果出现呼吸困难/脸唇肿胀/全身迅速扩散，请立刻急诊"
                    navigation:
                      hospitals:
                        - rank: 1
                          name: "北京协和医院"
                          is_3a: true
                          distance_m: 1200
                          reason: "三甲综合医院，距离较近，急诊/门诊齐全"
                        - rank: 2
                          name: "中日友好医院"
                          is_3a: true
                          distance_m: 3500
                          reason: "三甲综合医院，口碑较好"
                        - rank: 3
                          name: "朝阳医院"
                          is_3a: false
                          distance_m: 900
                          reason: "距离更近，可作为备选"
                      route_plan:
                        mode: "driving"
                        eta_min: 15
                        summary: "大约15分钟车程"
                    weather:
                      summary: "小雨，气温5°C"
                      tips:
                        - "带伞出行"
                        - "注意保暖"
                    disclaimer: "本建议仅供参考，不替代专业医疗诊断"
                    hotline_tip: "如果你不确定症状严重程度，或者情况在变重，建议你也可以拨打当地医疗热线或直接拨打医院电话先确认"
                emergency_triage:
                  summary: EMERGENCY triage (skips evidence retrieval)
                  value:
                    status: "final"
                    session_id: "550e8400-e29b-41d4-a716-446655440001"
                    turn_count: 1
                    triage_level: "EMERGENCY"
                    recommended_departments: ["急诊"]
                    possible_causes:
                      - "可能的心脏问题（疑似）"
                      - "呼吸系统问题（疑似）"
                    self_care_tips:
                      - "保持坐姿，减少活动"
                      - "如有硝酸甘油，按医嘱使用"
                    red_flags:
                      - "症状紧急，建议立即急诊/呼叫急救"
                    navigation:
                      hospitals:
                        - rank: 1
                          name: "北京协和医院"
                          is_3a: true
                          distance_m: 800
                          reason: "三甲综合医院，距离最近，有急诊"
                        - rank: 2
                          name: "朝阳医院"
                          is_3a: false
                          distance_m: 500
                          reason: "备选项"
                        - rank: 3
                          name: "中日友好医院"
                          is_3a: true
                          distance_m: 2000
                          reason: "三甲备选"
                      route_plan:
                        mode: "driving"
                        eta_min: 8
                        summary: "大约8分钟车程"
                    disclaimer: "本建议仅供参考，不替代专业医疗诊断"
                    hotline_tip: "如果你不确定症状严重程度，或者情况在变重，建议你也可以拨打当地医疗热线或直接拨打医院电话先确认"
                need_clarification:
                  summary: Insufficient information, need clarification
                  value:
                    status: "need_more_info"
                    session_id: "550e8400-e29b-41d4-a716-446655440002"
                    turn_count: 1
                    clarify_questions:
                      - "头痛持续多久?"
                      - "是否伴有发热或恶心?"
                      - "头痛部位具体在哪里?"
                    disclaimer: "本建议仅供参考，不替代专业医疗诊断"
                    hotline_tip: "如果你不确定症状严重程度，或者情况在变重，建议你也可以拨打当地医疗热线或直接拨打医院电话先确认"
                external_service_degraded:
                  summary: External service failure (graceful degradation)
                  value:
                    status: "final"
                    session_id: "550e8400-e29b-41d4-a716-446655440003"
                    turn_count: 1
                    triage_level: "ROUTINE"
                    recommended_departments: ["内科"]
                    possible_causes:
                      - "消化系统问题（疑似）"
                    self_care_tips:
                      - "清淡饮食"
                      - "多休息"
                    red_flags:
                      - "如症状加重或出现呕吐、黑便，请立即急诊"
                    navigation:
                      hospitals: []  # Empty: Amap service failed
                      route_plan: null
                    disclaimer: "本建议仅供参考，不替代专业医疗诊断"
                    hotline_tip: "如果你不确定症状严重程度，或者情况在变重，建议你也可以拨打当地医疗热线或直接拨打医院电话先确认"

        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  error_message:
                    type: string
                    example: "text exceeds 2000 character limit"
              required:
                - status
                - error_message

        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  error_message:
                    type: string
                    example: "Unexpected system error"
              required:
                - status
                - error_message

        '503':
          description: Service temporarily unavailable (external service degradation)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  error_message:
                    type: string
                    example: "Partial service degradation - navigation unavailable"
              required:
                - status
                - error_message

  /health:
    get:
      tags:
        - triage
      summary: Health check endpoint
      description: Returns service health status for load balancers
      operationId: triage.health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  version:
                    type: string
                    example: "1.0.0"
                  timestamp:
                    type: string
                    format: date-time
                required:
                  - status
                  - version
                  - timestamp

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        **Note**: Authentication is OUT OF SCOPE for v1.0 (per OUT-012).

        Future versions may add API key authentication for rate limiting.

security: []

x-amazon-apigateway-request-validator:
  body:
    model: Empty
    required: true
  requestParameters:
    querystring:
      required: false
    header:
      required: false